feat(hub-common): add datasetToContent() and datasetToItem()
affects: @esri/hub-common, @esri/hub-content

moved from hub-content, which continues to re-export them until the next breaking change
