+++ packages/common/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "abab": "^2.0.5",
     "adlib": "^3.0.7",
+    "jsonapi-typescript": "^0.1.3",
     "tslib": "^1.13.0"
   },
   "peerDependencies": {
